refactor: split git tests into unit and integration
This commit doesn't change what is being tested. It only moves the
git integration tests from the unit test suite to the integration test
suite.
